python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
deepface==0.0.79
opencv-python==4.8.0.76
Pillow==10.0.0
numpy==1.24.3
requests==2.31.0
scipy==1.11.2
scikit-learn==1.3.1
